+++ mbgdml/predictors/gap_predict.py
@@ -27,8 +27,7 @@ log = logging.getLogger(__name__)
 
 # Possible ray task.
 def predict_gap(
-    z, r, entity_ids, entity_combs, model, periodic_cell, ignore_criteria=False,
-    **kwargs
+    z, r, entity_ids, entity_combs, model, periodic_cell, **kwargs
 ):
     """Predict total :math:`n`-body energy and forces of a single structure.
 
@@ -48,9 +47,6 @@ def predict_gap(
         GAP model containing all information need to make predictions.
     periodic_cell : :obj:`mbgdml.periodic.Cell`, default: ``None``
         Use periodic boundary conditions defined by this object.
-    ignore_criteria : :obj:`bool`, default: ``False``
-        Ignore any criteria for predictions; i.e., all :math:`n`-body
-        structures will be predicted.
     
     Returns
     -------
@@ -96,11 +92,9 @@ def predict_gap(
                 continue
         
         # Checks criteria cutoff if present and desired.
-        if model.criteria_cutoff is not None and not ignore_criteria:
-            _, crit_val = model.criteria_desc_func(
-                z_comp, r_comp, None, entity_ids[r_slice]
-            )
-            if crit_val >= model.criteria_cutoff:
+        if model.criteria is not None:
+            accept_r, _ = model.criteria.accept(z_comp, r_comp)
+            if not accept_r:
                 # Do not include this contribution.
                 continue
         
@@ -116,7 +110,7 @@ def predict_gap(
     return E, F
 
 def predict_gap_decomp(
-    z, r, entity_ids, entity_combs, model, ignore_criteria=False, **kwargs
+    z, r, entity_ids, entity_combs, model, **kwargs
 ):
     """Predict all :math:`n`-body energies and forces of a single structure.
 
@@ -174,17 +168,16 @@ def predict_gap_decomp(
         
         z_comp = z[r_slice]
         r_comp = r[r_slice]
+        
         if first_r:
             atoms = ase.Atoms(z_comp)
             atoms.set_calculator(model.gap)
             first_r = False
         
         # Checks criteria cutoff if present and desired.
-        if model.criteria_cutoff is not None and not ignore_criteria:
-            _, crit_val = model.criteria_desc_func(
-                z_comp, r_comp, None, entity_ids[r_slice]
-            )
-            if crit_val >= model.criteria_cutoff:
+        if model.criteria is not None:
+            accept_r, _ = model.criteria.accept(z_comp, r_comp)
+            if not accept_r:
                 # Do not include this contribution.
                 continue