+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- ``mbgdml.descriptors.Criteria`` class for setting up structure descriptors and cutoffs for models.
 - Model analysis with Matérn covariance function.
 - ``mbe_contrib`` tests.
 - Radial distribution function analysis
@@ -14,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
+- ``mbgdml.criteria`` is now ``mbgdml.descriptors`` with a modular handling of descriptors and cutoffs.
 - Provide changeable ``Z``, ``R``, ``E``, and ``F`` keys for loading ``npz`` data sets.
 - Modularize the error and loss calculations.
 - Reduce regularization strength to ``1e-10`` instead of ``1e-15``.